[nfc] Merge refactors from upstream trunk
