+++ openmp/libomptarget/deviceRTLs/amdgcn/src/libcall.cu
@@ -13,6 +13,7 @@
 
 #include "omptarget-nvptx.h"
 #include "target_impl.h"
+#include "device_environment.h"
 
 // Timer precision is 1ns
 #define TIMER_PRECISION ((double)1E-9)
@@ -291,7 +292,7 @@ EXTERN void omp_get_partition_place_nums(int *place_nums) {
 }
 
 EXTERN int omp_get_cancellation(void) {
-  int rc = FALSE; // currently false only
+  int rc = 0;
   PRINT(LD_IO, "call omp_get_cancellation() returns %d\n", rc);
   return rc;
 }
