+++ openmp/libomptarget/deviceRTLs/amdgcn/src/omptarget-nvptx.h
@@ -15,21 +15,14 @@
 #define __OMPTARGET_NVPTX_H
 
 // std includes
-#include <stdint.h>
+#include <inttypes.h>
+#include <math.h>
 #include <stdlib.h>
 
-#include <inttypes.h>
-
-// cuda includes
-#ifndef __AMDGCN__
-#include <cuda.h>
-#endif
-#include <math.h>
-
 // local includes
+#include "target_impl.h"
 #include "debug.h"     // debug
 #include "interface.h" // interfaces with omp, compiler, and user
-#include "option.h"    // choices we have
 #include "state-queue.h"
 #include "support.h"
 
@@ -99,33 +92,6 @@ private:
 
 extern __device__ __shared__ omptarget_nvptx_SharedArgs
     omptarget_nvptx_globalArgs;
-
-// Data sharing related quantities, need to match what is used in the compiler.
-#ifdef __AMDGCN__
-enum DATA_SHARING_SIZES {
-  // The maximum number of workers in a kernel.
-  DS_Max_Worker_Threads = 960,
-  // The size reserved for data in a shared memory slot.
-  DS_Slot_Size = 256,
-  // The slot size that should be reserved for a working warp.
-  DS_Worker_Warp_Slot_Size = WARPSIZE * DS_Slot_Size,
-  // The maximum number of warps in use
-  DS_Max_Warp_Number = 16,
-};
-#else
-enum DATA_SHARING_SIZES {
-  // The maximum number of workers in a kernel.
-  DS_Max_Worker_Threads = 992,
-  // The size reserved for data in a shared memory slot.
-  DS_Slot_Size = 256,
-  // The slot size that should be reserved for a working warp.
-  DS_Worker_Warp_Slot_Size = WARPSIZE * DS_Slot_Size,
-  // The maximum number of warps in use
-  DS_Max_Warp_Number = 32,
-  // The size of the preallocated shared memory buffer per team
-  DS_Shared_Memory_Size = 128,
-};
-#endif
 
 // Data structure to keep in shared memory that traces the current slot, stack,
 // and frame pointer as well as the active threads that didn't exit the current
@@ -370,8 +336,6 @@ private:
   uint64_t cnt;
 };
 
-#include "../../../src/device_env_struct.h"
-
 /// Memory manager for statically allocated memory.
 class omptarget_nvptx_SimpleMemoryManager {
 private:
@@ -387,12 +351,6 @@ public:
   INLINE void Release();
   INLINE const void *Acquire(const void *buf, size_t size);
 };
-
-////////////////////////////////////////////////////////////////////////////////
-// global device envrionment
-////////////////////////////////////////////////////////////////////////////////
-
-extern __device__ omptarget_device_environmentTy omptarget_device_environment;
 
 ////////////////////////////////////////////////////////////////////////////////
