+++ openmp/libomptarget/deviceRTLs/amdgcn/src/supporti.h
@@ -267,26 +267,6 @@ INLINE void *SafeFree(void *ptr, const char *msg) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Named Barrier Routines
-////////////////////////////////////////////////////////////////////////////////
-
-#ifdef __AMDGCN__
-INLINE void named_sync(const int barrier, const int num_threads) {
-  // we have protected the master warp from releasing from its barrier
-  // due to a full workgroup barrier in the middle of a work function.
-  // So it is ok to issue a full workgroup barrier here.
-  __builtin_amdgcn_s_barrier();
-}
-#else
-INLINE void named_sync(const int barrier, const int num_threads) {
-  asm volatile("bar.sync %0, %1;"
-               :
-               : "r"(barrier), "r"(num_threads)
-               : "memory");
-}
-#endif
-
-////////////////////////////////////////////////////////////////////////////////
 // Teams Reduction Scratchpad Helpers
 ////////////////////////////////////////////////////////////////////////////////
