+++ openmp/libomptarget/deviceRTLs/interface.h
@@ -17,6 +17,7 @@
 #ifndef _INTERFACES_H_
 #define _INTERFACES_H_
 
+#include <stddef.h>
 #include <stdint.h>
 
 #ifdef __AMDGCN__
